Create Chronometer.cs
This is my first try at creating a chronometer program. It basically does the job of a regular chronometer with functions such as 
Start - starts the chronometer by declaring one variable with data type DateTime.Now as the time of start, 
Stop - finds the difference between two recorded variables and prints the time difference,
Lap - this method declares another variable as DateTime.Now and finds the difference between the stop variable and the start variable and also uses a List<string> to write the  last lap's duration and prints the recorded time,
Laps - is method that returns the list of recorded laps with indexation and {minutes}:{seconds}:{milliseconds} formatting,
GetTime - returns the last recorded time,
Reset - sets all the counters and clears the lap memory.